WIP: Creating new Alert Troubleshooting section (JSC#667)
A customer has requested that we provide further detail on
the alerts and how to manage from a customer perspective.

This is a WIP on a new troubleshooting section specifically
created for this purpose. The alerts were previously placed in
an appendix, but have since been removed from that appendix.