Creating new Alert Troubleshooting section (JSC#667) (#149)
* WIP: Creating new Alert Troubleshooting section (JSC#667)

A customer has requested that we provide further detail on
the alerts and how to manage from a customer perspective.

This is a WIP on a new troubleshooting section specifically
created for this purpose. The alerts were previously placed in
an appendix, but have since been removed from that appendix.

* Removing from book file (JSC#667)

* Committing half of the alerts so I don't lose them (JSC#667)

* Committing again so I don't lose more work (JSC#667)

* Last of all the commits (JSC#667)

* Multiple updates to jsc#SES-667

* inserted prompts
* replaced <literal>'s with <option>'s where applicable
* applied entities where applicable
* capitalized contents of <replaceable>'s
* aligned Ceph version names to capitalized
* few general fixes

Co-authored-by: tbazant <tbazant@users.noreply.github.com>