Proposing removal of cepham prompt to ceph
We currently use the ceph-salt tool (meaning the commands are `ceph-salt BLAH` to deploy our Ceph cluster instead of the upstream `cephadm bootstrap`. I propose removing the mention of cephadm in the entities in favor of simply using `ceph`. The `cephadm` command is used a total of once in the deployment guide, and I believe this is misleading. We don't want customers to confuse the options, and think they can operate the SES cluster with Ceph commands. If you use the upstream commands, you will get an upstream container instead, and the cluster will be unsupported.